requirements.txt
requests==2.31.0
python-dotenv==1.0.1
spacy==3.7.2
pydantic==2.5.3
vk-api==11.9.9
fastapi==0.109.0
uvicorn==0.27.0
openai==1.55.3
google-generativeai==0.8.3
aiogram==3.13.1
psycopg2-binary==2.9.9
flask==3.0.0
markdown2==2.4.12
